requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-SocketIO
python-socketio
python-engineio
Werkzeug
python-dotenv
Flask-JWT-Extended
openai
eventlet
